Deduplicate HolderRef FromStr into holder_ref! macro, rename UtxoRef to OutPoint
